Add Node.js OS Network Interfaces
Open to re-formatting this, as it felt a bit much to list every one of the affected libraries when the core issue is known, but it's the biggest reason I haven't been able to make much use of Bash on Windows. Almost any Node.js library that creates a local server fails until the fix for Microsoft/BashOnWindows#468 comes down the pipe.